requirements.txt
aiogram>=3.26.0
beautifulsoup4>=4.14.3
fastapi>=0.115.0
uvicorn>=0.34.0
python-dotenv>=1.0.0
requests>=2.32.5
# playwright>=1.58.0
# selenium>=4.41.0
# webdriver-manager>=4.0.2